Strip out HTML tags in token replacement
Template tags in the HTML body in a message can include HTML tags, which e.g. happens when copying column titles from the spreadsheet over to the composer in Gmail. This change strips out HTML from template keys.

Also changed the regex for matching {{...}}, so that it used a capturing group for the key.